+++ framework-system/framework-management/src/main/java/uk/gov/justice/services/management/shuttering/command/ShutterSystemCommand.java
@@ -5,8 +5,9 @@ import uk.gov.justice.services.jmx.command.BaseSystemCommand;
 public class ShutterSystemCommand extends BaseSystemCommand {
 
     public static final String SHUTTER_APPLICATION = "SHUTTER_APPLICATION";
+    private static final String DESCRIPTION = "Shutters the application to allow for maintenance.";
 
     public ShutterSystemCommand() {
-        super(SHUTTER_APPLICATION);
+        super(SHUTTER_APPLICATION, DESCRIPTION);
     }
 }