+++ framework-utilities/src/main/java/uk/gov/justice/services/framework/utilities/cdi/CdiProvider.java
@@ -1,0 +1,10 @@
+package uk.gov.justice.services.framework.utilities.cdi;
+
+import javax.enterprise.inject.spi.CDI;
+
+public class CdiProvider {
+
+    public CDI<Object> getCdi() {
+        return CDI.current();
+    }
+}