+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/command/SystemCommand.java
@@ -5,5 +5,5 @@ import java.io.Serializable;
 public interface SystemCommand extends Serializable {
 
     String getName();
-
+    String getDescription();
 }