+++ jmx/jmx-command-handling/src/test/java/uk/gov/justice/services/jmx/command/TestCommand.java
@@ -5,6 +5,6 @@ public class TestCommand extends BaseSystemCommand {
     public static final String TEST_COMMAND = "TEST_COMMAND";
 
     public TestCommand() {
-        super(TEST_COMMAND);
+        super(TEST_COMMAND, "This is a command used for testing");
     }
 }